RE-1767 Ensure retries cleanup
The allocate_pubcloud playbook will try in a second region
if creating an instance in the first region fails, or if the instance
in the first region can't be contacted via ssh.

This commit ensures that any instance created in the first region is
removed before attempting to create an instance in a second region.

This commit also ensures that writing the inventory file is the first
task after each instance creation task so that the cleanup stage has
the highest chance of success.

This commit also switches back from wait_for_connection to wait_for
as wait_for_connection ignores host and port parameters, only checking
connectivity to the host the task is targeted at.