Use the script module for apt sources reconfiguration
Similar to the raw module, the script module does not
require python on the target host.

When using the raw module, the output of the apt sources
file has whitespace prefixed in the resulting file and
therefore contains ' EOF' at the end of it, which apt
find to be a misconfiguration and fails.

We therefore switch to using a script instead which is
copied over and executed by the script module.

JIRA: RE-1767