Create pluggable artifact upload hook
In this PR, we create a pluggable artifact upload hook which
currently implements two types: 'log' and 'file'.

This new hook makes use of the in-repo ``component_metadata.yml``
file to discover its configuration.

The 'log' type works in exactly the same way as the current job
artifact publishing, and has a shim for the
``RE_HOOK_ARTIFACT_DIR`` environment variable for backwards
compatibility.

The 'file' type provides the capability to upload any file or
recursively upload any folder to public cloud's cloudfiles, in
a container name which matches the repository running the test.

With the pluggable mechanism in place, we can switch the
default mechanism used for all jobs to use it. The shim
in place to cater for a repository which does not have
any component_metadata will cater for uploading logs
as it did before.

JIRA: RE-1325