Improve jenkins slave setup resilience
The current play combines the apt cache update and the
package install task in one, which is nice and concise
but makes it difficult to understand whether the install
failed due to an apt cache update failure or a package
install failure. Compounding the issue is the fact that
the apt module swallows the error in Ansible 2.3, so we
do not get to see it. Also, by default the apt_repository
module tries to update the apt cache after changing the
repository configuration, but if that fails then it will
retry and find on the second round that the file needs no
change and will not bother doing the cache update.

To improve our ability to diagnose errors when executing
these tasks, we move the setup of apt debug mode to be
done on all instances created (not just those without
python). We then disable the cache update when the
repository has changed, and implement a subsequent apt
cache update task on its own. We also convert the single
apt module task into two seperate shell tasks which each
do one thing. This means if either fail we have a specific
output of the error in more detail.

We also ensure that the retries/delays are consistent
with those recently implemented for the OnMetal nodes.

JIRA: RE-1783