RE-1490 Remove unneeded rpco snapshot jobs
We're currently building newton/newton-rc snapshots for
trusty/trusty_loose_artifacts and xenial/xenial_loose_artifacts but
as only 'loose' artifact mode is being used in production we should
force all jobs to use the 'loose' image and skip building the 'strict'
images.

This commit adds a separate project for newton/newton-rc for snapshots
so we can be specific about which images get built.

Additionally, we bump all existing jobs using snapshots to use the
'loose' newton image, and bump all other snapshots to the latest
released version. Note that for queens, we actually fix the image to
rpc-r17 rather than rpc-r18.