RE-1490 Add job to bump snapshot versions
This commit adds a new job called Bump-Snapshot-Images that cycles
through rpc-openstack branches finding the latest tag and then bumps
versions of snapshot images in rpc_gating/.

Note that before we update any images, we test the image to ensure that
it is functional. Currently the latest snapshot image for each branch
will be tested regardless of whether a change has been detected, but
this serves as a bit rot test to help us identify problematic
rpc-openstack snapshots before users do.