RE-1787 Fix missing PR link in dep update JIRA issue
The pull request link is sometimes missing from the dependency update
JIRA ticket (see RE-1725) because in the case when dep_update.sh creates
a PR, it updates the JIRA issue message prior to creating the PR, thus
not yet having a PR link.  This change modifies dep_update.sh to use the
correct order and create/extract the PR link.

JIRA: RE-1787