Extend release to publish details to Confluence
The purpose of this change is to automate the publishing of releases to
the Confluence wiki.

`release.py` is extended to include the new sub-command
`publish_release_to_wiki`. This updates the table on the page to include
an entry for the new release. If the release already exists on the page,
its details, such as comment, are updated.

If a release is done within 7 days of the first working day of the month
it is considered a scheduled release, if it is outside of that window it
is considered an asynchronous release. The wiki page "Product Releases"
details the differences between these two types of release.

JIRA: RE-1534