RE-1839 Build Summary Date Accounting Errors
Addresses two grave bugs in the way build summary handles
dates when producing graphs:

1) The graphs are generated using a histogram, builds are divided into
60 24 hour buckets, and each bucket is labeled with a date. The problem
is that the start point is whenever the page is rendered so the 24 hour
histogram buckets do not line up with calendar days. The resolution is
to move the histogram start point to a day boundary so that histogram
buckets and days lineup.

2) The way that builds are allocated to buckets is by working out the
difference between the histogram end, and the build timestamp in
milliseconds, this in the divided by the number of milliseconds in a day
and rounded to produce an integer. That integer is subtracted from the
length of the histogram to get the bucket number.  Due to standard
rounding some builds would be put in the correct bucket, and some in the
bucket for the next day. The fix is to use the ceil function instead of
round.