Skip release tests with existing successful builds
`testRelease` is extended to enable it to discover any existing RELEASE_
or PM_ build that completed successfully for a particular SHA. RELEASE_
and PM_ jobs are treated as equivalent if they only differ by the prefix
(RELEASE_/PM_). If a build exists, the job does not need to be run
again.

To determine if a test was for a particular SHA the parameter
`_BUILD_SHA` is checked.

This change should, in general, reduce the time the release process
takes to complete, while allowing it to automatically resolve any
missing prerequisites.

JIRA: RE-1671