Add _BUILD_SHA parameter for tracking changes
`_BUILD_SHA` is added to simplify the discovery of completed builds that
were used to test a specific commit. A lot of jobs test a branch by
default which makes it difficult to determine exactly what was tested
without searching the log.

`_BUILD_SHA` is set once the project checkout is complete. It is set to
the current HEAD. For merged code this will mean it is set to the tip of
the branch being tested, in the case of a pull request this means that the
merge SHA (pr/ID/merge) will be used.

JIRA: RE-1671