Implement our own ansible.cfg at the highest precedence
The various tests which run may include implementing an ansible.cfg
file in the default search paths that ansible uses, resulting in
our playbooks trying to load/use things which are not meant to be.

To combat this, we ensure that we have our own ansible.cfg file and
that we use the environment variable to inform ansible where it should
find it so that it does not search other paths.

We also move ANSIBLE_HOST_KEY_CHECKING and ANSIBLE_FORCE_COLOR into
ansible.cfg so that the settings are easier to find in one place.

JIRA: RE-1870