Override default ansible plugin paths to prevent interference
The various tests which run may include implementing an ansible.cfg
file in the default search paths that ansible uses, resulting in
our playbooks trying to load/use things which are not meant to be.

To combat this, we override the paths for all plugin types and set
them to the location in the rpc-gating venv. We do the same for the
roles packaged into the rpc-gating tarball.

JIRA: RE-1870