Fix failure due to use of dir with Docker slave
The current version of Jenkins and Docker in use do not support
switching directories with `dir` inside a Docker slave. This change
addresses the issue by having `clone_with_pr_refs` return the SHA-1 hash
for the commit that is currently checked out. This removes the need to
have to discover the hash at a later point in the code.

JIRA: RE-1671