Ensure merge requirements met before gate/release
This change seeks to make gate and release trigger builds fail early if
GitHub branch protection requirements are not met. This helps to protect
against builds with side effects failing on merge. Specifically,
Component-Gate-Trigger_ and Component-Release-Trigger will now fail if
either the required number of approvals have not been given or the
required checks have not passed.

The new function `github.is_pr_approved` allows for GitHub checks to be
excluded from the test so that, for example, a build triggered by a
required GitHub check can make use of the function even though it will
not yet be registered as succeeding by GitHub.

JIRA: RE-1630