Update trigger phrase for gate and release
This change standardises the trigger phrase for the gate and release
process. :shipit: as a phrase is shorter and hopefully easier to
remember and also corresponds to an emoji that is commonly used on
GitHub to indicate approval of a change. By having one common phrase,
this should also reduce the impact of its use on developers.

JIRA: RE-1568