RE-1923 Cryptography version update
This commit bumps cryptography as there are problems with the wheels
for the version we were using previously.

It also adds OS detection to the lint script and sets appropriate
vars so that c libs from homebrew can be located when building
wheels. This shouldn't affect builds in CI (due to the conditional) but
will enable developers to build venvs (and run lint) on local
macOS environments.