RE-1599 Add dep update for dependents after release
This change triggers any component that is dependent on a project to be
(dep) updated immediately after the project is released instead of
relying solely on a weekly dep update.  In this configuration, the third
party dependencies are not updated.  The dependent search is performed
via modifications also made in the component tool (rpc_component).

Requested changes include:
* reading yaml format instead of tokenized text.

JIRA: RE-1599