Enable releasing with failed tests
Sometimes projects choose to release with failed tests. While this is
not ideal, to cater for this requirement, this pull request adds the
ability to skip code validation. Normally what happens is that if a
project has tests that have not been performed for the specified hash,
the release process will look to run those tests to ensure the code
works. With this change the status of any existing builds are ignored
and no new builds are started. All tests are skipped and the rest of the
release is completed as if the code was validated.

To skip validation a release should be triggered with the phrase
`:shipit: skip validation`.

JIRA: RE-1948