Enable configurable status context prefix
The GitHub status context used be standard jobs is prefixed with `CIT/`.
This is a legacy convention from when there were multiple Jenkins
instances and was used to distinguish between them. Removing or changing
this now would impact any repository with required checks.

This change allows the prefix to be modified for individual jobs while
maintaining the original as a default. This is being done to allow
rcbops/releases to have more descriptive checks.