Add component metadata file validation
component_metadata.yml can include information related to dependencies
and artefacts. To ensure this file is valid this change updates the
release pre-merge job, which is automatically added to new components,
to also test component metadata to ensure it is always in a valid
format.

The Docker file updated is used by the release job, python3 is installed
to support the component CLI tool.

JIRA: RE-1811