Add RPCO snapshot support to Nodepool
nodepool.yml is updated to support three new types of node. Replicating
how the snapshots are currently used, each node type is flavour 7 (15GB
Standard Instance). The nodes are labelled based on the major series to
which they belong and not the specific version, this simplifies the
management of the configuration but does mean if a project requires use
of a specific version it will need to be done manually.

DFW pool boot-timeout increased to 30 minutes to given snapshots a
greater chance of building. Due to the longer time to build, the
min-ready, max-ready-age and max-servers values have been increased,
compared to standard images, to reduce the chance that users are
affected by the longer build time.

bump_snapshots.yml has been updated so that Nodepool is automatically
updated when new snapshots become available.

Executing the thaw is moved to ensure it works for Nodepool nodes in
addition to those still building the single use slaves manually.

JIRA: RE-1465