Reset python in rpc-gating venv at run time
After extracting the venv on a target host, the python version
on the host may be different enough to cause the scripts used
in the venv to fail. This patch ensures that the python version
in the venv is replaced with whatever is on the host at runtime.

Resetting python in the venv cannot be done by simply doing a
python binary copy - it results in a broken venv which is missing
some of the standard libraries. Due to that, we need to use the
'virtualenv' tool for the reset. Given the range of options
available over time, we have to do a fair amount of logic to
figure out the options available to use.

To help prove that this works, unit tests are added for the
following:

1. Get a nodepool bionic node.
2. Publish artifacts using a xenial container (instead of using
   the long-lived slave), and a xenial nodepool node.
3. Publish artifacts using a bionic container, and a bionic
   nodepool node.
4. Verify that at least one script is working using the venv when
   building any single-use slaves (container, nodepool, and
   jenkins-built).

The current standard job Dockerfile installs a lot of content into
it, so the environment is not pristine. The package 'groovy2' is
also not available in Ubuntu Bionic. As part of this patch, we
reduce what the Dockerfile installs to a minimal set to support
running tests and executing the rpc-gating artifact uploading.

Note that the use of the rpc-gating scripts in a container is not
guaranteed to work. The artifact uploads work, but the jira-utils
script does not. The issue is that the venv python replacement is
done on the long-running-slave (Xenial), but then trying to use it
in the Bionic (or other) platform in the container fails because
the standard libraries aren't right for that distro. As such, I've
switched the Container Custom Args test to just use Xenial.

JIRA: RE-1977