RE-1522 Modify Grafana graphs that may have aggregation
Some of the graphs in Grafana appear to have aggregation and/or sparse
data, making the plots very difficult to understand.  After examining
the Graphite data and also comparing with the Openstack nodepool data,
our data/graphs appears to be correct but the panel configurations need
to be adjusted.

JIRA: RE-1522