Expose the build result to the post stage
To ensure that standard jobs are able to recognise whether the run
stage passed/failed, we export a new environment variable before
running the post script. This makes it possible to adjust the post
script to skip the artifact configuration which causes broken images
to get uploaded after the job.

JIRA: RE-1989