RE-1980 Fix the Grafana time range issue
Some of the Grafana graphs show inaccurate data points if the graph
is zoomed out, or the time range is larger than about 6-12 hours.
For more an example, see the notes file attached to the RE-1980 card.
This task attempts to resolve the time range issue and includes a
minor adjustment to fix the units (scale) in the "Time to Ready" panel.

Ref: https://github.com/rcbops/rpc-gating/blob/master/CONTRIBUTING.md

JIRA: RE-1980