Adjust nodepool settings to give more headroom to DFW
The public cloud DFW region is exclusively used for snapshot
builds and snapshot consumption, both of which currently use
jenkins (not nodepool). In the 2018.10 release we'll be migrating
the snapshot consumption to use nodepool, so that will help us to
remain within a quota - but jenkins does not have quota awareness,
so we have the snapshot creation jobs failing unless we provide
some headroom for jenkins to use.

In this patch we do the following:

1. Reduce the DFW nodepool quota to ensure that there is headroom
   for jenkins.
2. Reduce the min-ready for ubuntu-xenial-g1-8 and ubuntu-bionic-g1-8
   because it seems that they are underutilised with the current
   setting. With min-ready at 1, there will be one available in each
   region which should be more than enough.
3. Reduce the min-ready for ubuntu-bionic-p2-15 because it is currently
   unused by any jobs.
4. Increase the max-servers for Phobos to be 30, which matches the
   setting for ORD.

JIRA: RE-2027