Ensure archive_artifacts reports failure in tests
`archive_artifacts` creates a Jira issue when it encounters a failure as
well as swallowing the exception, this prevent the archiving process
causing build failures. The act of suppressing the exception means that
failures when testing with RE-unit-test-artefact-publication are not
causing the builds to fail.

This change modifies `archive_artifacts` so that creating issues can be
disabled, this prevents rpc-gating pull request check failures from
needlessly creating new Jira issues. In addition, it is now possible to
toggle whether or not an exception is raised if archiving artefacts
fails.

JIRA: RE-1654