ASC-1042 Rename asc staging image
This commit removes the specialized naming convention for the images
used for the asc staging job. The non-standard name prevents the images
cached in the CDN from being detected.

In order to allow the job to remain discrete from the production system
testing job, the action is renamed to 'system_staging'.