RE-2064 Run install_ansible in containers
In standard_job_slave we call use_node for all node types apart from
containers. This means that install_ansible is not run on containers,
which means the rpc-gating venv is not adjusted for the current os.

This commit adds a call to install_ansible within each container before
its used.