Disable systemd-resolved in bionic+ images
Bionic images are failing to resolve DNS because glean
is not implementing any systemd-resolved configuration
to make it work. This results in the system initially
working OK, but failing if any changes are made to the
networking configuration which causes systemd-resolved
to restart (replacing the content of /etc/resolve.conf
which glean put there).

While this is currently only showing up in RPC-O tests,
the issue would happen on any nodepool nodes which make
adjustments to the network config. Given that we use
glean as part of the nodepool node setup, we need to do
this here to prevent test failures rather than expect
test implementations to do it.

We disable this on any image which is not trusty/xenial
so that this will happen for any images bionic onwards.

JIRA: RI-514