Configure systemd-resolved on bionic+ images
In Ubuntu Bionic onwards, systemd-resolved is enabled by
default.

We have glean to setup the interfaces and the resolvers
from config-drive. Ideally we'd like glean to configure
systemd-resolved properly, but it can't do that right
now.

We've tried the following:

1. Disabling systemd-resolved. This causes other services
   which are set to depend on it not to start, so the VM
   never boots.
2. Using resolvconf. This has the same problem as glean -
   systemd-resolved keeps changing the config to something
   broken.

So, as a workaround, we implement a static config here with
public DNS servers.

JIRA: RI-514