Update hardening tests to use MNAIO images
In order to ensure that the hardening tests make use of
MNAIO images rather than do the deployment from scratch,
we rename the 'action' to 'hardening' rather than 'deploy'.

We also update the nodepool node to use bionic.

Finally we change the 'image' axis to be 'xenial_mnaio'
for pike, queens rocky and master rather than make mention
of 'loose_artifacts' given that these RPC-O branches do not
have any notion of artifacts. We restrict that naming to
only newton. We also add bionic VM tests for rocky and
master.

JIRA: RE-2053