Add 'elk' action to RPC-O MNAIO tests
In order to test the ELK add-on for RPC-O, we add an
extra 'action' which executes that option using the
MNAIO images. This also ensures that the 'deploy'
action remains focused on the base RPC-O deployment
without any extras.

The 'action' axis components are all re-orders into
alphabetical order for easier maintenance.

JIRA: RI-263