Add artefact store support to release workflow
When the release job is run, on an rcbops/releases pull request to add
an artefact store, the change must be recognised so that the automated
merge can be done. This change adds the necessary support. rpc_component
is updated to a version that supports the `artifact-store` subcommand.

JIRA: RE-2100