Add LON region to nodepool
In order to provide additional capacity for tests using
OnMetal nodes, we add the LON region to nodepool and
configure it only for those specific labels. We also add
the graphing for the new region to grafana.

Given that the 'ubuntu-xenial-om-io2' label is no longer
in use, we also reduce the min-ready for that node type
to 0.

Finally, the LON region uses a different authentication
endpoint so we add a new 'cloud' to the clouds.yaml file
which provides the appropriate account credentials and
the right endpoint.

JIRA: RE-2152