RE-2087 Catch workspace cleanup failures
When we release a node, we clean the workspace by calling
deleteDir from common.use_node. If the node has gone offline
for any reason, that deleteDir will fail. If the node wen't offline
during the post phase, we don't want its disapperance to fail the
build so we enclose deleteDir in try/catch.