RE-2083 Allow constraints leak test during maint
A prefix match is used to determine which jobs can run during
a maintenance window. One of the existing prefixes is
RE-unit-test, which was designed to cover all the unit tests,
however the constraints leak job name starts RE-unit-venv.
This commit reduces the prefix match to RE-unit to ensure that the
constraints leak job is allowed.