Adds temporary job to capture older versions
Captures MNAIO environment after the initial deployment of a
version has ran.  Leverages the workarounds to install an older
version so that we can create a library of older versions and
focus on just testing upgrades rather than deployment of an older
verision.

Uses snapcap as the action which skips the upgrade part of
the job and captures the image on success. Testing out with one
image to see how it works.