RE-2061 Ignore non pr failures on maint tickets
At the end of a maintenance the mainteance-end job attempts to recheck
all PR builds that were aborted due to the maintenance. This process
fails if any non-pr builds are reported, as the PR related fields are
null.

This PR ensures that non-pr failures are safely skipped when rechecking
PRs after a maintenance.