Add RE_JOB_PROJECT_NAME environment variable
RE_JOB_NAME currently binds to the name of the JJB project and not the
name of the job. To enable this to be correct, this change adds the new
variable RE_JOB_PROJECT_NAME to allow existing uses of RE_JOB_NAME to be
switched to using RE_JOB_PROJECT_NAME before updating RE_JOB_NAME to the
correct value.

This change adds a new pre-merge test job for rpc-gating to test that
the `RE_JOB_.*` environment variables are defined as expected.

JIRA: RE-2031