Fix RE_JOB_NAME value
This change fixes RE_JOB_NAME to ensure it is bound to the name of the
job and not the name of the project. The job name does not appear to be
available when using jenkins-jobs and so this change makes use of the
environment variable JOB_NAME to set RE_JOB_NAME to the correct value.

JIRA: RE-2031