Reduce retention and leave logs uncompressed for 5 days
A change [1] in the workflow-job-plugin has resulted in jenkins
no longer being able to render compressed job logs. As such, we
need to disable compression. We are not able to expand our
available disk space for logs and cannot meet our current retention
without compression, so we have to reduce the retention and disable
the compression for anything younger than 5 days.

[1] https://github.com/jenkinsci/workflow-job-plugin/commit/46bf528f5965ecb94180d2a01d91bf7cb05da9f5
JIRA: RE-2215