Update snapshot image for RPC-O Queens
The bump snapshot job only does a search and replace for
N-1 to N where N is the most recent release. Due to the
fact that the last in-use images are N-2, the search and
replace is not working, even though the image is confirmed
to work.

As such, to get things back on track we handle this bump
manually.

JIRA: RI-538