RE-2267 Fix sources param default in std lint job
The jjb_params parameter macro requires an argument to set the default
value, however no argument was passed. Consequently the rendered
job contained the text "{JOB_SOURCES}" as the default value of that
paramater, rather than the expected value from defaults.
This commit fixes that by passing the appropriate argument.