RE-1204 Render jjb before extracting dsl for lint
Previously we extracted DSL blocks from jjb yaml and ran
the DSL blocks through the groovy syntax checker. This could cause
problems if there were jjb macros in the groovy code.

This commit alters the lint process so that the jjb files are
rendered to XML job definitions, then the DSL block is extracted
from the XML and checked as before.