Use nodepool images for AIO snapshot builds
Rather than use public cloud images, which have proven to be unreliable
due to the nova-agent not starting properly, we enable the jenkins node
build to use the images built by nodepool instead.

The playbook fetches the list of available images, then uses the latest
available one for the build.

This brings the image build process closer to what is used for all tests
and should be more reliable given that nodepool does not use nova-agent.

JIRA: RE-2278