Update requirements and constraints
In order to update to using Ansible 2.7, we need to include a
general update of most of the libraries to ensure compatibility.

We are able to remove the python-novaclient requirement, now that
Ansible no longer requires it.

To also ensure that we have the required SSL modules for successful
SNI connectivity to SSL sites, we change the 'requests' requirement
to 'requests[security]'.

JIRA: RE-1721