Implement a common method for rpc-gating runtime requirements
Currently there are multiple methods for preparing a host for
tests:

1. In the jenkins nodepool plugin configuration, there is a script
   which executes on all target nodes to prepare the node for use.
   This script is used to add the jenkins user, add the ssh keys,
   add the JDK, etc. This is used for preparing OnMetal/Ironic
   hosts where we do not control the image.

2. In the setup_openstack_instances playbook, we use the raw Ansible
   module to add python to the target host so that Ansible is able
   to execute other tasks and modules on that host.
   This is used for preparing VM's where we do not control the image.

3. In the allocate_pubcloud playbook we use the raw and script
   modules to ensure that the apt sources are configured, apt debug
   logging is enabled, and that python and python-yaml are installed
   on the target host.
   This is used for preparing VM's where we do not control the image.

In this patch, we combine these all into a single script which is
used as userdata provided to cloud-init [1], which is present in all
the images for public cloud and Phobos.

In order to use this for nodepool, we also need to update the nodepool
SHA we use to include the support for userdata. In the nodepool config
we unfortunately cannot import the file directly into the config, so
we use another method of getting the same file for the same SHA.

As part of this, we also make the following minor adjustments:

1. In the allocate_pubcloud playbook we use the Ansible playbook_dir
   magic variable, rather than refer to the WORKSPACE environment
   variable. This makes the playbook executable for testing without
   having to set the WORKSPACE environment variable.

2. In the user_data_publoud script, we remove the unattended upgrade
   disabling. Now that all RPC-O tests make use of nodepool images
   the issue which brought that content there is no longer relevant.

[1] https://cloud-init.io
JIRA: RE-1523