RE-2310 fix shouldrerelease logic
Previously this function used .matches() however that requires
the pattern to match the whole string. The intention was for
search behaviour, where true is returned if the pattern can be
found at any position in the input string.

.matches() was added to ensure the result is Boolean, but as the
method signature declare a Boolean result, the match object is
automatically casted to Boolean so the call is unnecessary.