Add retry logic to wiki release page updates
Confluence protects against unknowingly updating a page that has been
modified by using a page version ID. The ID supplied when an update is
requested must be exactly one larger than the current page version ID.
However this means that if there is a conflict, publishing the update
will fail causing the release build to fail. This change adds retry
logic so that any failures when attempting to update the wiki, with a
new version of the release page, will result in the process being rerun.
This means any new updates will be incorporated in the subsequent
attempt.

JIRA: RE-2327