Fix use of container with periodic clean-up
A recent change broken the building of images in the periodic clean-up.
This change fixes this by switching the job to separate the build
process from it being run.

Build-Summary-Docker-Build is replaced with a parameterised version
(Build-Docker-Images-For-Master) that can be used to build all Docker
images, defined in rpc-gating, that are required to be run on master.

The use of the Docker plugin is removed from the periodic clean-up job
because the plugin and installed version of Docker are not compatible.
This issue existed prior to the use of the registry but is fixed here to
ensure that the use of the container is as expected.

JIRA: RE-404