Improve management of registry errors
This change adds a timeout of 10 minutes when pushing an image,
generally only a few minutes should be required at most. This timeout
removes the need to wait 20 minutes or more for the push to time out.

The end of the final retry no longer attempts to recover the VPN
connection given there will be no more tries.

The wording of exception, raised when retries exhausted, is improved to
indicate where to look for more information.

JIRA: RE-404