RE-2355 Add internal registry creds to ETL job
This job requires access to the Jenkins API as it pulls data from there
to send to google big query.

The only nodes we have with access to the Jenkins api are the master
and one TES node. This node runs centos6, which is a pain as it
uses python2.6

This job uses a docker image to insulate it from the host, allow
modern python and package installs etc. In order to push/pull the image
from the repo, creds are required.