Use shared slave for addArtifactTypeToComponent
addArtifactTypeToComponent requires a number of C libraries to be
installed which causes errors if the slave does not have them. This
change moves addArtifactTypeToComponent to use a shared slave so that
project slaves do not have additional requirements placed on them given
the potential for conflict or absence.

JIRA: RE-2306