Prevent failure when removing non-existent slave
The legacy method of provisioning slaves using playbooks as part of the
build (not Nodepool) is still used in some cases. Part of the clean up
during the build is to delete the node from Jenkins. Removing the slave
from Jenkins when it does not exist causes a build failure. This change
updates the code to prevent false-positives from these failures causing
builds to fail or mask the true failure.

There are two changes made by this patch:
- The delete call in the Python script is wrapped with exception
  handling so that failures when deleting unknown nodes do not
  propagate.
- `ssh_slave.destroy` is used to delete the node from Jenkins and is now
  performed before the instance is destroyed so that it doesn't
  disappear from Jenkins while still registered and result in misleading
  logs.

JIRA: RE-2348