Ensure that glean sets up eth0 on boot of snapshots
When booting up a snapshot, there are no network interfaces present.
The 'glean' service should set them up automagically once the devices
are present, but it is not working. To work around this, we enable
the service ourselves before we shut the instance down. It will then
start on boot and the new instance should have network connectivity.

JIRA: RE-2278