Update newton snapshot image name to latest release
The snapshot bump job has been failing for some time due to issues
with the snapshots, and the job only caters for a single version
bump. RPC-O newton released several tags in quick succession, so
we need to manually bump this to the most recent release because
the job won't do it automatically.

JIRA: RE-2278