Ensure that interfaces setup by glean are removed
When booting up a snapshot, if there are previously implemented
interfaces present which were setup by glean, glean will skip
setting up new interfaces from config_drive. To ensure that the
new instance gets new interfaces setup, we force the removal of
the old interfaces.

For systemd environments, the clean-up happens automatically -
but for upstart environments that doesn't happen, so we need to
ensure it is forced.

JIRA: RE-2278