Reduce nodepool host name length
The current nodepool nodes have a host name which includes a lot
of unnecessary information. For example, the provider does not
need to be included in the host name. Also, some of the label
names are based on very long image names, and the label name is
also included in the host name.

This, in some circumstances, results in a host name which exceeds
the maximum of 63 characters. For example, the following is 67
characters:

nodepool-rpc-r18-xenial_no_artifacts-swift-pubcloud-dfw-0000152558

When trying to set this host name, glean fails because the operating
system rejects it. This causes glean to fail and not continue to
implement the network interfaces, resulting in connectivity for the
node never coming up.

To resolve this, we remove the unnecessary inclusion of the provider
in the host name and we also shorten the nodepool labels to only
include the necessary information to differentiate it. We also then
change all the jobs to make use of the correct new labels.

Finally, we add a missing entry to ensure that the label
'rpc-r14-trusty-swift' has a provider.

ref: https://en.wikipedia.org/wiki/Hostname#Restrictions_on_valid_hostnames
JIRA: RE-2278